% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audit_categorical_cleaning.R
\name{audit_categorical_cleaning}
\alias{audit_categorical_cleaning}
\title{Title}
\usage{
audit_categorical_cleaning(
  df,
  raw_col,
  cleaned_col,
  var_name = NULL,
  plot = TRUE
)
}
\description{
Description
}
