% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticate_google_drive.R
\name{authenticate_google_drive}
\alias{authenticate_google_drive}
\title{authenticate_google_drive Title}
\usage{
authenticate_google_drive()
}
\value{
TRUE if the authentication was successful
}
\description{
Authenticate with Google Drive using the appropriate credentials
The pipeline will not run if this step is not successful.
}
\examples{
authenticate_google_drive()
authenticate_google_drive()
}
