% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_age_group.R
\name{calculate_age_group}
\alias{calculate_age_group}
\title{Calculate Age Group from Date of Birth}
\usage{
calculate_age_group(dob, reference_date = Sys.Date())
}
\arguments{
\item{dob}{A vector of date-of-birth values (class Date)}

\item{reference_date}{A single date to calculate age relative to (defaults to today)}
}
\value{
A character vector of age groups
}
\description{
Calculate Age Group from Date of Birth
}
