% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_2018_vaccinations.R
\name{clean_2018_vaccinations}
\alias{clean_2018_vaccinations}
\title{Title}
\usage{
clean_2018_vaccinations(df)
}
\description{
Description
}
\examples{
health_data_ \%>\%
  clean_2018_vaccinations() -> health_data_
  
health_data_ \%>\%
  select(contains("vacc_")) \%>\%
  skimr::skim()
}
