% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_opencensus_dob.R
\name{clean_opencensus_dob}
\alias{clean_opencensus_dob}
\title{Clean OpenCensus SRP Date of Birth}
\usage{
clean_opencensus_dob(open_census_)
}
\value{
tibble
}
\description{
This function cleans the date of birth (DoB) data from the Open Census SRP dataset.
}
