% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_health_diagnoses.R
\name{create_health_diagnoses}
\alias{create_health_diagnoses}
\title{Title}
\usage{
create_health_diagnoses(df, diagnosis_lookup = malagasy_diagnosis_lookup_table)
}
\description{
Description
}
\examples{
health_data_ \%>\% 
  create_health_diagnoses() \%>\%
  summary()
}
