% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_to_local.R
\name{download_to_local}
\alias{download_to_local}
\title{download_to_local}
\usage{
download_to_local(f_regex, download_dir)
}
\value{
The path to the downloaded file
}
\description{
Download the google drive file to local storage
}
\examples{
authenticate_google_drive()
test_file <- "openSRP_dictionary.xlsx"
fpath <- download_to_local(test_file, here::here())
file.exists(fpath)
file.remove(fpath)
}
