% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_drive_file_changed.R
\name{has_drive_file_changed}
\alias{has_drive_file_changed}
\title{has_drive_file_changed}
\usage{
has_drive_file_changed(
  file_id,
  last_known_run_time = ymd_hms("2025-04-01 00:00:00", tz = "UTC")
)
}
\arguments{
\item{file_id}{The ID of the file to check}

\item{last_known_run_time}{The last known time the pipeline was frozen for running}
}
\value{
Boolean
}
\description{
Check from a file modification timestamp whether or not it has been changed
since last run of the pipeline.
}
\examples{
test_file <- "1X9pd4nOjl33zDFfTjw-_eFL7Qb9_g6VfVFDp1PPae94"
has_drive_file_changed(test_file, lubridate::ymd_hms("2025-04-01 00:00:00", tz = "UTC"))
}
