% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mahery_input_files.R
\name{mahery_input_files}
\alias{mahery_input_files}
\title{Track the Input Files for the Mahery Cohort}
\usage{
mahery_input_files(freeze_date = mdy("01-04-2025"))
}
\value{
tibble with the file names and their regex patterns
}
\description{
Description
}
