% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_2018_health.R
\name{preprocess_2018_health}
\alias{preprocess_2018_health}
\title{preprocess_2018_health Title}
\usage{
preprocess_2018_health()
}
\value{
1
}
\description{
preprocess_2018_health Title
}
\examples{
preprocess_2018_health()
}
