% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_2018_household_heads.R
\name{preprocess_2018_household_heads}
\alias{preprocess_2018_household_heads}
\title{preprocess_2018_household_heads Preprocess 2018 Household Heads}
\usage{
preprocess_2018_household_heads(opensrp, cohort_2018)
}
\value{
Tibble with cleaned household head data
}
\description{
preprocess_2018_household_heads Preprocess 2018 Household Heads
}
\examples{
preprocess_2018_household_heads(opensrp, cohort_2018) \%>\%
  summary()
}
