% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_opensrp2018.R
\name{preprocess_opensrp2018}
\alias{preprocess_opensrp2018}
\title{Title}
\usage{
preprocess_opensrp2018(open_census)
}
\description{
Description
}
\examples{
preprocess_opensrp2018(open_census) \%>\% summary()
}
