% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_rdp_health_subset2019.R
\name{preprocess_rdp_health_subset2019}
\alias{preprocess_rdp_health_subset2019}
\title{RDP Health Subset Preprocessing}
\usage{
preprocess_rdp_health_subset2019(dharma2019)
}
\description{
This function preprocesses the 2019 data to extract a subset of health-related information.
It focuses on cleaning and organizing health data for the RDP paper.
}
