% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_rdp_health_subset2020.R
\name{preprocess_rdp_health_subset2020}
\alias{preprocess_rdp_health_subset2020}
\title{Title}
\usage{
preprocess_rdp_health_subset2020(dharma2020, dharma2019_dict)
}
\description{
Description
}
\examples{
preprocess_rdp_health_subset2020(dharma2020, dharma2019$`Data Dictionary`) \%>\%
  skim()
}
