% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_uid.R
\name{generate_uid}
\alias{generate_uid}
\title{Title}
\usage{
generate_uid(salt = "cohort_secret", ..., nchar = 6)
}
\description{
Description
}
\examples{
generate_uid(salt = "cohort_secret", "John", "Doe", "1990-01-01")
}
